+++ zircon/kernel/object/include/object/fifo_dispatcher.h
@@ -46,11 +46,11 @@ class FifoDispatcher final : public PeeredDispatcher<FifoDispatcher, ZX_DEFAULT_
   ktl::variant<zx_status_t, UserCopyCaptureFaultsResult> ReadToUserLocked(
       size_t elem_size, user_out_ptr<uint8_t> ptr, size_t count, size_t* actual) TA_REQ(get_lock());
 
-  const uint32_t elem_count_;
+  const uint64_t elem_count_;
   const uint32_t elem_size_;
 
-  uint32_t head_ TA_GUARDED(get_lock());
-  uint32_t tail_ TA_GUARDED(get_lock());
+  uint64_t head_ TA_GUARDED(get_lock());
+  uint64_t tail_ TA_GUARDED(get_lock());
   ktl::unique_ptr<uint8_t[]> data_ TA_GUARDED(get_lock());
 
   static constexpr uint32_t kMaxSizeBytes = ZX_FIFO_MAX_SIZE_BYTES;
